LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.quadraticCurveTo.nonfinite-expected.txt

2d.path.quadraticCurveTo.nonfinite
quadraticCurveTo() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS quadraticCurveTo() with Infinity/NaN is ignored
